[Feat] 검색 유사도에 따른 우선순위 정렬
