+++ Cote/Cote/Domain/Usecase/CreateFolderUseCase.swift
@@ -12,15 +12,15 @@ protocol CreateFolderUseCase {
 }
 
 struct DefaultCreateFolderUseCase: CreateFolderUseCase {
-    private let repository: NoteRepository
+    private let repository: NoteRepositoryProtocol
 
-    init(repository: NoteRepository) {
+    init(repository: NoteRepositoryProtocol) {
         self.repository = repository
     }
     
     @MainActor
     init() {
-        self.init(repository: RealmNoteRepository())
+        self.init(repository: NoteRepository())
     }
 
     func execute(name: String, parentID: UUID?) async throws -> UUID {
